Move more classes to the engine
