+++ scijava/scijava-ops-engine/src/main/java/org/scijava/ops/engine/BaseOpHints.java
@@ -1,4 +1,4 @@
-package org.scijava.ops.api.features;
+package org.scijava.ops.engine;
 
 import java.lang.annotation.Annotation;
