+++ scijava/scijava-ops-engine/src/main/java/org/scijava/ops/engine/MatchingConditions.java
@@ -1,5 +1,5 @@
 
-package org.scijava.ops.api.features;
+package org.scijava.ops.engine;
 
 import java.util.Objects;
