+++ scijava/scijava-ops-engine/src/main/java/org/scijava/ops/engine/impl/DefaultOpEnvironment.java
@@ -66,12 +66,12 @@ import org.scijava.ops.api.OpRef;
 import org.scijava.ops.api.OpRetrievalException;
 import org.scijava.ops.api.OpWrapper;
 import org.scijava.ops.api.RichOp;
-import org.scijava.ops.api.features.BaseOpHints.Adaptation;
-import org.scijava.ops.api.features.BaseOpHints.DependencyMatching;
-import org.scijava.ops.api.features.BaseOpHints.Simplification;
-import org.scijava.ops.api.features.MatchingConditions;
 import org.scijava.ops.engine.DependencyMatchingException;
+import org.scijava.ops.engine.MatchingConditions;
 import org.scijava.ops.engine.OpCandidate;
+import org.scijava.ops.engine.BaseOpHints.Adaptation;
+import org.scijava.ops.engine.BaseOpHints.DependencyMatching;
+import org.scijava.ops.engine.BaseOpHints.Simplification;
 import org.scijava.ops.engine.matcher.MatchingRoutine;
 import org.scijava.ops.engine.matcher.OpMatcher;
 import org.scijava.ops.engine.matcher.impl.DefaultOpMatcher;
