+++ scijava/scijava-ops-engine/src/main/java/org/scijava/ops/engine/matcher/MatchingRoutine.java
@@ -3,7 +3,7 @@ package org.scijava.ops.engine.matcher;
 
 import org.scijava.ops.api.OpEnvironment;
 import org.scijava.ops.api.OpRetrievalException;
-import org.scijava.ops.api.features.MatchingConditions;
+import org.scijava.ops.engine.MatchingConditions;
 import org.scijava.ops.engine.OpCandidate;
 
 /**
