+++ scijava/scijava-ops-engine/src/main/java/org/scijava/ops/engine/matcher/adapt/AdaptationMatchingRoutine.java
@@ -22,9 +22,9 @@ import org.scijava.ops.api.OpDependencyMember;
 import org.scijava.ops.api.OpEnvironment;
 import org.scijava.ops.api.OpInfo;
 import org.scijava.ops.api.OpRef;
-import org.scijava.ops.api.features.BaseOpHints.Adaptation;
+import org.scijava.ops.engine.BaseOpHints.Adaptation;
 import org.scijava.ops.engine.DependencyMatchingException;
-import org.scijava.ops.api.features.MatchingConditions;
+import org.scijava.ops.engine.MatchingConditions;
 import org.scijava.ops.engine.matcher.MatchingRoutine;
 import org.scijava.ops.engine.matcher.OpMatcher;
 import org.scijava.ops.engine.matcher.impl.DefaultOpRef;
